skip tests for py3 from running by python2
